+++ NEWS.md
@@ -11,7 +11,9 @@ NEW FEATURES :
 * `createClusterST()`/`editClusterST()` :  
   - **New dimension of time series** (all matrices will be of (8760, N), noting that N >= 1)
   - **Suppression of properties** (refreshtimeseries,refreshintervalload, refreshintervalhydro, refreshintervalwind, refreshintervalthermal, refreshintervalsolar from settings/generaldata.ini)
-  
+* `createCluster()` :   
+ - **Dynamic groups** (The group parameter is now dynamic and has no restrictions. The default group value for Renewable depends on the version: Other RES 1 if < 930, Other if >= 930).
+
 # antaresEditObject 0.9.2.9000
 (cf. Antares v9.2 changelog)