+++ NEWS.md
@@ -15,6 +15,7 @@ NEW FEATURES :
  - **Dynamic groups** (The group parameter is now dynamic and has no restrictions. The default group value for Renewable depends on the version: Other RES 1 if < 930, Other if >= 930).
 * `updateScenarioBuilder()` New types of series "sts" and "sta" ("sct apports level" and "sct contraintes") is available
 * `createDistrict()` uses a specific endpoint to create the district
+* `editDistrict()` allows the user to edit a district
 
 
 # antaresEditObject 0.9.2.9000