+++ NEWS.md
@@ -1,7 +1,7 @@
 > Copyright © 2016 RTE Reseau de transport d’electricite
 
 
-# antaresEditObject 0.9.3.9000
+# antaresEditObject 0.9.3
 (cf. Antares v9.3 changelog)
 
 NEW FEATURES :